Main Ansible code
Configure machines based on the tag we set in terraform.

The app tier installs Docker and run the latest of the
saliceti/hello-world image from Docker Hub.

The web tier installs Nginx and configures it as a proxy to the app
servers. The private IPs are retrieved from hostvars.
Unfortunately the module leaves Nginx default config so we remove it
manually. As a consequence, the playbook may not work the first time.
This should be investigate further.

Last is a very basic integration test. The public DNS name of the web
server is retrieved, then it is queried twice and we check the response
is different. This proves the stack works and we have round robin.